fix(pose_analysis): hip alignment
Hip alignment did not work some times. Specifically on the XZ plane. Not sure why. Replacing "pi + angle" with "- angle" in the rotation matrix calculation fixed the issue, but now the pose was upside down. This in turn could be fixed by removing the flip step later. Hip seems to be properly aligned now.