feat(app): go to body-detail screen from overview
Tapping the puppet on the result overview screen now takes the user to the body-part detail screen.

Caveat: There is no native way to only make non-transparent parts of an image clickable. Because of this, the top-most image in the puppet image stack is always hit. That means, whatever you tap you get send to the detail screen for the upper arm, because the top graphic is the right hand. This needs to be fixed using a custom hit-tester