feat: optimize puppet display load time
Until now each body-part image had to be loaded twice. Once for the pixel values to be used in the hit-detection and once as an asset for display. This change uses the `Image.memory` widget to allow us to use the already loaded bytes. This should hopefully speed up load time. Also refactor code a bit