feat: enhance body-part group type
Type is now a more complex ADT instead of a simple enum. This has the primary goal of encoding left/right differences into the type.

From the outside it should still be usable mostly as before.

This change also moves some functions which were previously unbound onto the BodyPartGroup type or into extensions.

This change also has the nice effect that the body-part detail screen now displays the correct score for the picked side of the body part, instead of a merged one.

This is in preparation for #98