refactor: body-part group score api
The function for getting body-part scores now returns a list instead of a single value. For singular body-parts the list will contain the single score and for paired body-parts it will contain left and right score. This way clients can choose how to combine scores themselves. Scores will also not be normalized inside the function but by callers, again offering flexibility in how the function is used.