refactor: use file-based session storage
This change introduces a file-based session storage model instead of the database we had before.

This done for two reasons:

1. The object-box code was somewhat complicated to read and edit, but this is actually a minor concern.
2. More importantly, the data we store does not suite itself to relational databases such as object-box. We have a bunch of stuff which more or less equates to int[] which can be stored far better in a simple text file than a table format.

This design will also work better for the keyframes we plan to store as part of sessions soon. These will likely be images which are trivial to store as files, but likely more difficult to store in object-box.