refactor: use provider (#115)
The idiomatic way of injecting dependencies in Flutter seems to be provider. I originally wanted to manage dependencies myself, because I don't like "magic" libraries were dependencies "teleport" around the app, but in this case I see how I am fighting the framework.

So this change adds the provider dependency and reworks the injection of the rula-session-repo to use it.