refactor: use widget based navigation (#116)
Seems like Flutter recommends to not use path based nativation.

So this change refactors the project to use simple widget based navigation instead. This had the benefit of making passing arguments easier. On the other hand screens now need to directly reference each other, which is tighter coupling.

In total I think this is an improvement.