refactor: onboarding decision
Instead of hardcoding the logic to always go to onboarding inside the welcome screen, we instead introduce the class OnboardingState which will be used to get the onboarding state from some persistent store. For now we use a ConstantOnboardingState which always says we have not yet done onboarding. So this is a refactor with no actual effect, but it brings us closer to where we want to go