feat: remove record button tutorial
We have a page in the pre-record tutorial popup which instructs users to press the 'Start' button. This is a leftover from when the button was just an icon and people did not know to press it. But now the button is a text button anyway so this instruction is not necessary.

This change removes that page and related localization text.