refactor: activity usage
Currently most screens use the Activity type in String form, ie they use the localized name of the activity. This is spreading UI concerns all throughout the app. Better to leave the activity in it's natural enum form as long as possible and only convert it to String at the point of presentation