fix: cut-off body graphic
The body graphic was cut off horizontally because it was scaled to fit the space vertically. But if we set it to BoxFit.contain instead, which fixes the issue, the graphic becomes pretty small. This is because the actual images had a lot of whitespace around them. This change removes that whitespace thus fixing the issue.

Resolves #211