fix: camera controller disposed error
Very often when we go from the analysis screen to the result screen there is a very brief flash of a red error screen. The app still functions as intended and goes to the result screen after a moment. The slower the device, the longer the error is visible.

This happens on the analysis screen, because for a moment the camera controller is disposed but we have not yet transitioned to the next screen.

This was now fixed by changing the timing of when the camera controller is disposed vs. when it is set to none. We now set it to none first, which prevents it from being accessed in build and then only afterwards dispose it. This gets rid of the red error and simply renders an empty black space.