fix: handle no-keyframes on result screen
Currently we get an exception when trying to view a session which has no keyframes. This can happen if the session is very static, so not enough change happened to trigger a keyframe.

We now make the vertical highlight optional and only display it if there is a keyframe.

Furthermore we only display the keyframe carousel if there is at least 1 keyframe.